Move lines of code an cyclomatic complexity to metrics module
