+++ detekt-core/build.gradle.kts
@@ -1,5 +1,6 @@
 dependencies {
     api(project(":detekt-api"))
+    api(project(":detekt-metrics"))
 
     testImplementation(project(":detekt-rules"))
     testImplementation(project(":detekt-test"))
